Rename test resources to helpers for consistency
