+++ tests/chopper_test_helpers.py
