+++ tests/conftest.py
@@ -1,1 +1,1 @@
-from tests.chopper_test_resources import chopper_details  # noqa: F401
+from tests.chopper_test_helpers import chopper_details  # noqa: F401
