+++ tests/test_disk_chopper_geometry_creator.py
@@ -9,7 +9,7 @@ from nexus_constructor.geometry.disk_chopper.disk_chopper_geometry_creator impor
     DiskChopperGeometryCreator,
     RESOLUTION,
 )
-from tests.chopper_test_resources import EXPECTED_Z
+from tests.chopper_test_helpers import EXPECTED_Z
 
 POINT_X = 2.0
 POINT_Y = 3.0
